Remove Iterable requirement
The requirement for a Graph to be iterable greatly reduces the ease of
use of the implementation on objects that are not traditional Graphs,
like multidimentional lists.

The iterable requirement was only utilized to validate the nodes. This
patch removes this validation and leaves it up to the caller to validate.
This in turn removes the need for the Graph to be iterable.

Signed-off-by: Rodrigo Chiossi <rodrigochiossi@gmail.com>